refactor: move common PlayerService code to AbstractPlayerService
