+++ app/src/main/java/com/github/libretube/services/OfflinePlayerService.kt
@@ -1,37 +1,18 @@
 package com.github.libretube.services
 
-import android.content.BroadcastReceiver
-import android.content.Context
 import android.content.Intent
-import android.content.IntentFilter
 import android.os.IBinder
-import androidx.annotation.OptIn
-import androidx.core.app.NotificationCompat
-import androidx.core.app.ServiceCompat
-import androidx.core.content.ContextCompat
-import androidx.lifecycle.LifecycleService
 import androidx.lifecycle.lifecycleScope
-import androidx.media3.common.C
 import androidx.media3.common.MediaItem
 import androidx.media3.common.Player
 import androidx.media3.common.util.UnstableApi
-import androidx.media3.exoplayer.ExoPlayer
-import androidx.media3.exoplayer.trackselection.DefaultTrackSelector
-import com.github.libretube.LibreTubeApp.Companion.PLAYER_CHANNEL_NAME
-import com.github.libretube.R
 import com.github.libretube.constants.IntentData
 import com.github.libretube.db.DatabaseHolder
 import com.github.libretube.db.obj.DownloadWithItems
 import com.github.libretube.enums.FileType
-import com.github.libretube.enums.NotificationId
-import com.github.libretube.enums.PlayerEvent
-import com.github.libretube.extensions.serializableExtra
 import com.github.libretube.extensions.toAndroidUri
-import com.github.libretube.extensions.updateParameters
 import com.github.libretube.helpers.PlayerHelper
 import com.github.libretube.obj.PlayerNotificationData
-import com.github.libretube.util.NowPlayingNotification
-import com.github.libretube.util.PauseableTimer
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
@@ -40,131 +21,41 @@ import kotlin.io.path.exists
 /**
  * A service to play downloaded audio in the background
  */
-class OfflinePlayerService : LifecycleService() {
-    private var player: ExoPlayer? = null
-    private var nowPlayingNotification: NowPlayingNotification? = null
-    private lateinit var videoId: String
+@UnstableApi
+class OfflinePlayerService : AbstractPlayerService() {
     private var downloadsWithItems: List<DownloadWithItems> = emptyList()
 
-    private val watchPositionTimer = PauseableTimer(
-        onTick = this::saveWatchPosition,
-        delayMillis = PlayerHelper.WATCH_POSITION_TIMER_DELAY_MS
-    )
-
-    private val playerListener = object : Player.Listener {
-        override fun onIsPlayingChanged(isPlaying: Boolean) {
-            super.onIsPlayingChanged(isPlaying)
-
-            // Start or pause watch position timer
-            if (isPlaying) {
-                watchPositionTimer.resume()
-            } else {
-                watchPositionTimer.pause()
-            }
+    override suspend fun onServiceCreated(intent: Intent) {
+        downloadsWithItems = withContext(Dispatchers.IO) {
+            DatabaseHolder.Database.downloadDao().getAll()
+        }
+        if (downloadsWithItems.isEmpty()) {
+            onDestroy()
+            return
         }
 
-        override fun onPlaybackStateChanged(playbackState: Int) {
-            super.onPlaybackStateChanged(playbackState)
+        val videoId = intent.getStringExtra(IntentData.videoId)
 
-            // automatically go to the next video/audio when the current one ended
-            if (playbackState == Player.STATE_ENDED) {
-                val currentIndex = downloadsWithItems.indexOfFirst { it.download.videoId == videoId }
-                downloadsWithItems.getOrNull(currentIndex + 1)?.let {
-                    this@OfflinePlayerService.videoId = it.download.videoId
-                    startAudioPlayer(it)
-                }
-            }
-        }
-    }
-
-    private val playerActionReceiver = object : BroadcastReceiver() {
-        override fun onReceive(context: Context, intent: Intent) {
-            val event = intent.serializableExtra<PlayerEvent>(PlayerHelper.CONTROL_TYPE) ?: return
-            val player = player ?: return
-
-            if (PlayerHelper.handlePlayerAction(player, event)) return
-
-            when (event) {
-                PlayerEvent.Stop -> onDestroy()
-                else -> Unit
-            }
-        }
-    }
-
-    override fun onCreate() {
-        super.onCreate()
-
-        val notification = NotificationCompat.Builder(this, PLAYER_CHANNEL_NAME)
-            .setContentTitle(getString(R.string.app_name))
-            .setContentText(getString(R.string.playingOnBackground))
-            .setSmallIcon(R.drawable.ic_launcher_lockscreen)
-            .build()
-
-        startForeground(NotificationId.PLAYER_PLAYBACK.id, notification)
-
-        ContextCompat.registerReceiver(
-            this,
-            playerActionReceiver,
-            IntentFilter(PlayerHelper.getIntentActionName(this)),
-            ContextCompat.RECEIVER_NOT_EXPORTED
-        )
-    }
-
-    override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
-        lifecycleScope.launch {
-            downloadsWithItems = withContext(Dispatchers.IO) {
-                DatabaseHolder.Database.downloadDao().getAll()
-            }
-            if (downloadsWithItems.isEmpty()) {
-                onDestroy()
-                return@launch
-            }
-
-            val videoId = intent?.getStringExtra(IntentData.videoId)
-
-            val downloadToPlay = if (videoId == null) {
-                downloadsWithItems = downloadsWithItems.shuffled()
-                downloadsWithItems.first()
-            } else {
-                downloadsWithItems.first { it.download.videoId == videoId }
-            }
-
-            this@OfflinePlayerService.videoId = downloadToPlay.download.videoId
-
-            createPlayerAndNotification()
-
-            // destroy the service if there was no success playing the selected audio/video
-            if (!startAudioPlayer(downloadToPlay)) onDestroy()
+        val downloadToPlay = if (videoId == null) {
+            downloadsWithItems = downloadsWithItems.shuffled()
+            downloadsWithItems.first()
+        } else {
+            downloadsWithItems.first { it.download.videoId == videoId }
         }
 
-        return super.onStartCommand(intent, flags, startId)
-    }
-
-    @OptIn(UnstableApi::class)
-    private fun createPlayerAndNotification() {
-        val trackSelector = DefaultTrackSelector(this@OfflinePlayerService)
-        trackSelector.updateParameters {
-            setTrackTypeDisabled(C.TRACK_TYPE_VIDEO, true)
-        }
-
-        player = PlayerHelper.createPlayer(this@OfflinePlayerService, trackSelector, true)
-        // prevent android from putting LibreTube to sleep when locked
-        player!!.setWakeMode(C.WAKE_MODE_LOCAL)
-        player!!.addListener(playerListener)
-
-        nowPlayingNotification = NowPlayingNotification(
-            this,
-            player!!,
-            NowPlayingNotification.Companion.NowPlayingNotificationType.AUDIO_OFFLINE
-        )
+        this@OfflinePlayerService.videoId = downloadToPlay.download.videoId
     }
 
     /**
      * Attempt to start an audio player with the given download items
-     * @param downloadWithItems The database download to play from
-     * @return whether starting the audio player succeeded
      */
-    private fun startAudioPlayer(downloadWithItems: DownloadWithItems): Boolean {
+    override suspend fun startPlaybackAndUpdateNotification() {
+        val downloadWithItems = downloadsWithItems.firstOrNull { it.download.videoId == videoId }
+        if (downloadWithItems == null) {
+            stopSelf()
+            return
+        }
+
         val notificationData = PlayerNotificationData(
             title = downloadWithItems.download.title,
             uploaderName = downloadWithItems.download.uploader,
@@ -176,7 +67,11 @@ class OfflinePlayerService : LifecycleService() {
             .firstOrNull { it.type == FileType.AUDIO }
             ?: // in some rare cases, video files can contain audio
             downloadWithItems.downloadItems.firstOrNull { it.type == FileType.VIDEO }
-            ?: return false
+
+        if (audioItem == null) {
+            stopSelf()
+            return
+        }
 
         val mediaItem = MediaItem.Builder()
             .setUri(audioItem.path.toAndroidUri())
@@ -191,37 +86,6 @@ class OfflinePlayerService : LifecycleService() {
                 player?.seekTo(it)
             }
         }
-
-        return true
-    }
-
-    private fun saveWatchPosition() {
-        if (!PlayerHelper.watchPositionsVideo) return
-
-        player?.let { PlayerHelper.saveWatchPosition(it, videoId) }
-    }
-
-    override fun onDestroy() {
-        saveWatchPosition()
-
-        nowPlayingNotification?.destroySelf()
-        nowPlayingNotification = null
-        watchPositionTimer.destroy()
-
-        runCatching {
-            player?.stop()
-            player?.release()
-        }
-        player = null
-
-        runCatching {
-            unregisterReceiver(playerActionReceiver)
-        }
-
-        ServiceCompat.stopForeground(this, ServiceCompat.STOP_FOREGROUND_REMOVE)
-        stopSelf()
-
-        super.onDestroy()
     }
 
     override fun onBind(intent: Intent): IBinder? {
@@ -235,5 +99,19 @@ class OfflinePlayerService : LifecycleService() {
     override fun onTaskRemoved(rootIntent: Intent?) {
         super.onTaskRemoved(rootIntent)
         onDestroy()
+    }
+
+    override fun onPlaybackStateChanged(playbackState: Int) {
+        // automatically go to the next video/audio when the current one ended
+        if (playbackState == Player.STATE_ENDED) {
+            val currentIndex = downloadsWithItems.indexOfFirst { it.download.videoId == videoId }
+            downloadsWithItems.getOrNull(currentIndex + 1)?.let {
+                this@OfflinePlayerService.videoId = it.download.videoId
+
+                lifecycleScope.launch {
+                    startPlaybackAndUpdateNotification()
+                }
+            }
+        }
     }
 }
